requirements.txt
appdirs==1.4.2
Django==1.10.5
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.5.4
mongoengine==0.11.0
olefile==0.44
packaging==16.8
pymongo==3.4.0
pyparsing==2.2.0
six==1.10.0
